fix(creation): empty/error states + feed RPC type cast + settings copy
- experience mode no longer hangs on "Loading…" when the feed is empty or
  errors: show a real empty/error state with a Back button. This was the
  "loads forever" after both boot and linking — experience is the default
  screen and the feed was returning [].
- creation_feed_for_user: cast q.qscore::numeric in the cold-start path. The
  quality view's qscore is double precision but the function returns numeric,
  so Postgres raised "structure of query does not match function result type"
  and the feed came back empty. (Re-run this migration in Supabase.)
- settings: drop the literal "/creation" path from the Browse-on-R1 copy.

+++ migrations/2025-06-18_creation_feed.sql
@@ -104,7 +104,9 @@ BEGIN
   IF NOT v_has_likes THEN
     -- Cold start: global quality, excluding already-seen.
     RETURN QUERY
-      SELECT q.id, q.qscore AS score, 'quality_fallback'::text
+      -- qscore is double precision (exp/ln in the view); the function returns
+      -- numeric, so cast or Postgres raises "structure of query does not match".
+      SELECT q.id, q.qscore::numeric AS score, 'quality_fallback'::text
       FROM creation_quality_scores q
       WHERE NOT EXISTS (
         SELECT 1 FROM store_feed_seen s
